extract users top data fetch into custom hook
